Bug Fixes!
Response messages look a bit different.

Override system does not throw CurrentModificationException anymore.

All options/arguments are now case-insensitive, that is, if you type "yellow" instead of "YELLOW" it will work perfectly as expected.

When a birthday date was overriden, the previous bar would keep/blink for a second and then update. (Fixed)